Update to skip if dead/high latency
Update on Parallel Ping to test a few pings first to see if host is alive. If it is not alive it reports and skips any additional work.

+++ Ping test parallel.ps1
@@ -15,9 +15,48 @@ $Custom1 = Import-Csv $CSV | ForEach-Object {[PSCustomObject]@{
 #Shorten Custom1 Object
 $Custom2 = $Custom1.IPAddress
 
+#Short or Long test
+Write-Host "
+
+Options:
+
+1. Short test (50 pings)
+2. Long test (250 pings)
+"
+do {
+    $ShortOrLong = Read-Host -Prompt "Select an option"
+    if ($ShortOrLong -notin 1..2) {
+        Write-Warning "Please pick a valid option"
+    }
+}
+while ($ShortOrLong -notin 1..2)
+Write-Host ""
+
 #Parallel ping. Results recorded to each appropriate file
-$Custom2 | ForEach-Object -ThrottleLimit 20 -Verbose -Parallel {
-    ping -n 50 $_ | Out-File -FilePath $using:OutLocation\$_.txt -Force -Append
+if ($ShortOrLong -eq 1) {
+    Write-Host "Running short ping test..."
+    $Custom2 | ForEach-Object -ThrottleLimit 20 -Verbose -Parallel {
+        $AliveTest = Test-Connection -Count 3 -Ping -IPv4 -DontFragment -TargetName $_
+        if ($AliveTest.status -notcontains "Success") {
+            Write-Warning "Dead host at $_ or high latency. Skipping"
+        }
+        else {
+            ping -n 50 $_ | Out-File -FilePath $using:OutLocation\$_.txt -Force -Append    
+        }
+    }
+}
+
+if ($ShortOrLong -eq 2) {
+    Write-Host "Running long ping test..."
+    $Custom2 | ForEach-Object -ThrottleLimit 20 -Verbose -Parallel {
+        $AliveTest = Test-Connection -Count 3 -Ping -IPv4 -DontFragment -TargetName $_
+        if ($AliveTest.status -notcontains "Success") {
+            Write-Warning "Dead host at $_ or high latency. Skipping"
+        }
+        else {
+            ping -n 250 $_ | Out-File -FilePath $using:OutLocation\$_.txt -Force -Append
+        }
+    }
 }
 
 #Location and name of results
